fix(anthropic): avoid cache token double counting
